Extract mapping tests
